move explicit insight logic into hook
