+++ web_frontend/src/App.tsx
@@ -5,15 +5,17 @@ import { useUiUpdateBackendPoll } from "./hooks/useUiUpdateBackendPoll";
 import { generateRandomUserId, setUserIdAndDeviceId } from "./utils/utils";
 import Cookies from "js-cookie";
 import MainLayout from "./layouts/MainLayout";
+import { useHandleExplicitInsights } from "./hooks/useHandleExplicitInsights";
 
 export default function App() {
   useTranscription();
   useUiUpdateBackendPoll();
+  useHandleExplicitInsights();
 
   useEffect(() => {
-    let search = window.location.search;
-    let params = new URLSearchParams(search);
-    let userId = params.get('userId');
+    const search = window.location.search;
+    const params = new URLSearchParams(search);
+    let userId = params.get("userId");
 
     if (userId == undefined || userId == null || userId == "") {
       console.log("No userID in URL - checking for existing userID");
