+++ web_frontend/src/components/ExplicitCard.tsx
@@ -1,47 +1,15 @@
 import { Text, useMantineTheme } from "@mantine/core";
 import { AGENT_ICON_NAMES, AGENT_ICON_PATHS, AgentName } from "../types";
 import CardWrapper from "./CardWrapper";
-import { uniqueId } from "lodash";
-import { useEffect } from "react";
-import { useRecoilState, useSetRecoilState } from "recoil";
-import {
-  entitiesState,
-  explicitInsightsState,
-  isExplicitListeningState,
-} from "../recoil";
+import { useRecoilValue } from "recoil";
+import { explicitInsightsState } from "../recoil";
 
 const ExplicitCard = () => {
   const theme = useMantineTheme();
-  const [explicitInsights, setExplicitInsights] = useRecoilState(
-    explicitInsightsState
-  );
-  const setEntities = useSetRecoilState(entitiesState);
+  const explicitInsights = useRecoilValue(explicitInsightsState);
   const lastEntity = explicitInsights.at(-1);
   const queryString = `Query: ${lastEntity?.query}`;
   const answerString = `Answer: ${lastEntity?.insight}`;
-  const setIsExplicitListening = useSetRecoilState(isExplicitListeningState);
-
-  useEffect(() => {
-    if (lastEntity?.insight) {
-      setEntities((prevEntities) => [
-        ...prevEntities,
-        {
-          uuid: uniqueId(),
-          agent_insight: `${queryString}\n${answerString}`,
-          agent_name: AgentName.COMMAND,
-        },
-      ]);
-      setIsExplicitListening(false);
-      setExplicitInsights([]);
-    }
-  }, [
-    answerString,
-    lastEntity?.insight,
-    queryString,
-    setEntities,
-    setExplicitInsights,
-    setIsExplicitListening,
-  ]);
 
   return (
     <CardWrapper
